refactor(share): unify legacy and v03 share flow core
